+++ modules/v26_2_R1/src/main/java/com/nisovin/shopkeepers/compat/v26_2_R1/package-info.java
@@ -1,0 +1,2 @@
+@org.eclipse.jdt.annotation.NonNullByDefault
+package com.nisovin.shopkeepers.compat.v26_2_R1;